added prepend instrumentation fot HTTPClient
